Bugfixing round 2
We now compile error-free on MML again, but this time without
skipping half of all theorems because they have numbers in them.
Results:

bignum: 177
linear: 168590
success: 1820821

So we are at 91% checking. Notably we have achieved 100% of schematizer
proofs. What remains are a handful of proofs that need bigger bignums,
and linear arithmetic.