bigger bignums
I tried to just use 64 bits but unfortunately this fails on only
one proof in the MML. So we do a proper bignum implementation this time,
with a small num optimization because the vast majority of numbers
in the MML are small.

We also remove the error recovery for overflow, both because it is now
much less likely and also because treating them as opaque constants
on overflow is just a roundabout way to get proof failures.