change FlexAnd representation
I was planning to do this eventually anyway, but it turns out that the
existing representation actually has a soundness bug, in addition to
the sketchy unpacking code in `expand()`, because it is not true that
equality of the endpoints implies equality of the FlexAnd expressions.

For example, `P[1+1] /\ ... /\ P[n]` represents `P[x]` for `x in 1..=n`
but if you substitute `n = 2+2` then you get the expression
`P[1+1] /\ ... /\ P[2+2]`, which if re-parsed actually results in
`P[x+x]` for `x in 1..=2`. These two expressions are not equivalent
despite having the same endpoints and this can be used to prove false.