Transfer, part 4: cluster registrations
The big change here is because we need to remove duplicate attributes
in the prel/ files for verification because of mizar bugs, but we can't
just call `Attr::cmp` because we don't have an appropriate global
context. We could create one at some expense, but here we are gambling
that none of the really weird cases actually show up and a plain sort
without context is sufficient. (So far, it's holding up on the MML.)