make unify monotonic
Earlier changes have left us with one failure in the MML, so we have
to do something about this. This is a simple fix, but it unfortunately
comes with a small overhead since we are now considering more potential
instantiations. Still, not considering them was pretty much a bug in the
first place, leading to strange failures that make no sense from the
user perspective.