.miz parser: debugged on MML
It can now parse everything without failing, although the results
have not yet been tested for correctness (and cannot be until the
MSM pass is complete).

One slightly unnecessary change which required a patch on MML
was the removal of the form `scheme: stmt; end` using a `;` instead of
`proof` in `scheme: stmt proof end`. The former syntax makes no sense at
all seeing as everywhere else in the language a `;` after a statement
means a trivial proof to be done by the checker, and the `end` is
confusingly paired. It would not be hard to support, but it's a trivial
fix to remove the few cases that were using this variant syntax.