Name resolution: initial implementation, debugging
This is tested on MML in analyzer-quick mode, more modes will wait for the
next commit.

We have a few more MML patches here. The one in trees_a is due to
inconsistency in the Mizar quantifier lifting algorithm; the
discrepancies seem to be very rare so I'm okay with that.

The extra reservation in msafree5 is because Mizar seems to allow
resolving reservations which are dependencies of other reservations to a
redeclaration with the same name; this is confusing and rare enough that
a patch seems easier than the overhaul and complication of the
algorithm.