Upgrade to MML v. 8.1.12_5.74.1441
The verifier mostly works on the new MML out of the box, which is good,
although they seem to have come up with new and unusual ways to torture
the verifier in this edition and some of the new articles have managed
to beat the record for longest check time. Some profiling is needed, and
possibly some quadratic algorithms need to be revisited.

Fixes #1