shorten proofs in xprimet1.miz
This is our first major modification to the MML.
The proofs in article xprimet1 were very inefficient, and
it was simpler to fix the proofs rather than to try to make the overflow
behavior exactly line up with pascal mizar.

The original proofs were clearly automatically generated, and the fixes
in this commit are also automatically generated (by an ad-hoc script,
not included).