add term cache garbage collection
This is a very unsophisticated check, we just clear the cache when it
gets above a certain limit. The reasoning is that if the cache becomes
too big, then the cost of finding a constant in the cache (which
involves O(log N) comparisons to other constants in the cache) grows
larger than the cost of reconstructing the term and its type from
scratch (the cost of which is proportional to the size of various
internal data structures plus the size of the term). The limit used here
is somewhat on the high side, and was chosen based on when the cache
size starts to visibly slow down verification.