rewrite number04 to be faster
This is another major MML modification, which rewrites the core
computational part of number04 to use a different proof strategy that is
52% shorter and 19x faster to verify on mizar-rs, 11x faster on
`verifier`. In short, rather than calculating `a^n mod b` by working out
all the powers of `a` and then modding by `b`, we do the mod during the
power calculation, avoiding the need for big numbers.

Mizar can handle big numbers just fine, but the large number of numerals
and numeral expressions at global scope here pollutes both the term
cache and the infer const cache, leading to effectively quadratic
performance with the verifier getting slower and slower as it progresses
through the file. In the rewrite we put all the numbers inside many
little scopes, so there is no buildup of junk.

verifier: 3:59 <- 46:16
mizar-rs: 0:17 <-  5:26 + 7 ovfl. err

As a result, we are now back to 100% success on the MML.

Fixes #3