Minimize xprimes1,2 for massive gains
This is the easiest proof optimization I could possibly make and yet it
has huge wins on these files. The proof involved a series of steps of
the form

  now
    11 = 2*5 + 1; hence not 2 divides 11 by NAT_4:9;
    11 = 3*3 + 2; hence not 3 divides 11 by NAT_4:9;
    11 = 5*2 + 1; hence not 5 divides 11 by NAT_4:9;
    11 = 7*1 + 4; hence not 7 divides 11 by NAT_4:9;
  end;

proving that 11 does not divide any smaller prime, but this is silly
since one only needs to check primes up to the square root of 11
(meaning that 5 and 7 in this example are superfluous). Even more
surprisingly, *the lemma applied proves exactly this*: in this example
we apply a theorem that says "p*p <= k < 25 implies p = 2 or p = 3",
so literally all we have to do is delete the steps for 5 and 7 and the
proof still goes through. I would have thought that the Mizar tools
automatically minimize this sort of thing, but it's possible that the
use of "now" is messing things up.

In any case, I deleted a bunch of redundant lines, resulting in a 93%
reduction in xprimes1 proof lines and a 96% reduction in xprimes2.
xprimes1 check time is reduced 20x for verifier and 24.5x for mizar-rs.
The memory and disk usage for verifier was also alarmingly high, and I
had difficulty running it to completion. I had to kill the verifier run
for xprimes2, so the reduction is >28x for verifier and a whopping 69x
for mizar-rs.

xprimes1:
  lines: 315577 -> 20730
  verifier: 29:33    + 10.8G rss + 6.96G .xml
         ->  1:25    +  640M rss + 16.3M .xml
  mizar-rs:  2:08.5  + 30.3M rss
         ->  0:05.25 + 10.3M rss

xprimes2:
  lines: 453831 -> 15799
  verifier: killed at 15:54   + 12.1G rss + 16.9G .xml (out of disk)
         ->            0:34   +  586M rss + 18.4M .xml
  mizar-rs:            4:02   + 43.7M rss
         ->            0:03.5 + 10.7M rss